Refactor for safety
This is a large change that removes a significant amount of unsafe code.
The first step was to change PlainMem from `Rc<UnsafeCell<Vec<u8>>>` to
`Rc<RefCell<Vec<u8>>>` and properly use interior mutability. This had
some serious consequences:

Deref, as implemented in the MemView trait, is not possible to
implement. There is a discussion of this problem on
[this stack overflow post](https://stackoverflow.com/questions/57856047/how-can-i-implement-deref-for-a-struct-that-holds-an-rcrefcelltrait)

To work around this, an `as_deref()` method that uses a type of
`DerefReturn` was added to the `MemView` trait, and implemented.

The largest source of diff noise was a change to use the new
`as_deref()` method instead of either implicilty dereferencing
things or using a hard dereference operator. On the flip side,
ugly syntax like `&**rev.get_view(...)` turned into
`rev.get_view(...).as_deref()` which seemed a little more palatable.

Methods that mutate the state of the memory must use a
`&mut self` rather than `&self`. This affects methods like
`MemStore::write` and all the types that implement `MemStore`. There
was also a need to impelement `DerefMut` on the child classes so that
a mutable reference could be obtained; otherwise it's not possible
to call methods that mutate the base object.

A final small change was necessary to initialize the database by
getting unique ownership during a reset; see `firewood/src/db.rs`
for those changes.