Rename structs for better readability
These structures could use more meaningful names, making the code
a little easier to follow. Here are the renamed traits and structs
along with a brief description of each:

  1. MummyItem -> Storable
   This trait is used to indicate that the structure will be stored
   in a cached storage layer

  2. MummyObj -> StoredView
   This struct is implemented on a StorableStruct and saves the location
   where the item is actually stored

  3. MemStore -> CachedStore
   This trait has a few implementations:
    PlainMem: keeps all items in fixed sized memory
    DynamicMem: keeps all items in a growing memory
    StoreRevShared: read-only version of StoreRevMut
    StoreRevMut: the writeable revision store

  4. MemView -> CachedView
   Provides read-only access to a portion of a CachedStore

Only other change here was better enforcement for instantiation of
a StoredView; it now requires a StorableStruct (previously a MummyItem)
in order to instantiate it, rather than only when using some of the
methods, just to catch future compile errors earlier.