Attempt at fixing #45
Looks like the file being checked isn't the same as the file
being created so the test failed. This file does actually exist,
but for some reason the test was still just using /tmp for storing
the database. Updated the test and let's see if this passes now.

Also, since only one request can be in the queue at once, I did a
few get_page requests to see if that resolves the issue with timing.