Code cleanup: unused but constructed if expression
The if expression is never used, and a lot of work was done
to make sure everyone returns a stream. Just made the if
return () as expected.

The rle encoder docs are clear about the fact that you *may*
chain, and perhaps this code originally chained, but i doesn't
any more.