Code cleanup
I think we only need to get the counter from the metric to increment it,
so I removed the extra unused instructions. I'm a little concerned that
the counter being sent to the registry is not the same counter we are
holding in the hashmap.

I'm also a little concerned about lifetimes, since the registry and the
counter don't seem to have a lifetime relationship; I'm not sure what
prometheus would do if we increment the counter after destroying the
registry.

Removed unused test section as well.