Rename ObjPtr<T> to DiskAddress
The name "Object Pointer" implies an in-memory pointer, whereas what
it is actually being used for is a reference to an object stored on
disk

ObjPtr had a generic type T used to indicate the type of object stored
at that location. This had a lot of cognitive overhead and I didn't
find this all that useful, particularly since there were only a few
types of T that could be stored. Almost always this is a `Node`, but
it could also be a `CompactSpaceHeader` for free space within a file.

DiskAddress is implemented as a Option<NonZeroUsize>, which required
a partial change from u64 addresses to usize types. This could use
more work; it would be better to perhaps use DiskAddress types down
at the IO layers. I avoided this and stopped just short of that, adding
a bunch of `as usize` and `as u64` to avoid blowing up this diff too
much.

DiskAddress of 0 is invalid -- there is a header stored there that is
not used by the caching system. However, the code had some "bugs" where
it was reading unused data at address 0 (related to the blob deprecation
for eth) as well as some test cases that don't initialize the database
correctly, leading to objects at address 0. These were not addressed in
this fix, but left with TODOs to cause panics when operating at address
0.

Also noticed that merkle.rs uses Option<DiskAddress> which is (and
probably was) redundant, since DiskAddress(0) isn't valid for anything
that we want to cache. There is a file header there but it's handled
separately already.

The basics of this change were:
 - Change all ObjPtr<T> to DiskAddress
 - Change all ObjPtr::new_from_addr(a) to DiskAddress::from(a)
 - Change all ObjPtr::null() to DiskAddress::null()
 - Implement basic math operations on DiskAddress
   - std::ops::{Add, Sub, AddAssign, SubAssign} etc
   - allow adding/subtracting DiskAddress as well as usize types
 - Implement serialization and deserialization to/from [u8; 8]
   - added to_le_bytes() for serialization
   - added From<[u8; 8]> and From<&[u8]> for deserialization
 - Change ObjPtr::addr() to DiskAddress::get()
   - This should panic if DiskAddress is `None`, but doesn't currently