Remove Display implementation
All the use cases seem to be in the implementation of Debug for most
other types that embed this, with the exception of dump. I switched dump
to use Debug as well as it uses Debug for most other things it dumps
anyway and it seems to mostly be for debugging.