Write out a real CompactSpaceHeader early
During DB initialization, we end out using an unwritten
CompactSpaceHeader. This change writes one to disk when the database is
created, causing allocations of new objects to start at offset 1024
instead of at offset 0.

Writing objects at offset zero was working for a while because we have
an unused Node object there. That will get removed on the next commit.