Streaming iterator from start
This walks down the leftmost branch every time, recording the path down,
then figuring out what the key and value is at the bottom, then resumes.

Also refactors get_key_from_parents, as this is now done 3 times in the
code, sometimes when we add a leaf key value, and other times when not.