Enable lints for discussion
These lints are only in fwdctl because it was easy to fix the few errors
that popped up. Once we agree these are the right lints, we can move
them to the top level.

For ease of review, here are links to the lint descriptions we are
enabling as well reasons why I think they are important:

 - https://rust-lang.github.io/rust-clippy/rust-1.74.0/index.html#/unwrap_used
 - - code that panics should be avoided, except in tests
 - - not all panics have good error messages
 - https://rust-lang.github.io/rust-clippy/rust-1.74.0/index.html#/indexing_slicing
 - - panics
 - - there are good alternatives to index dereference that do not panic
 - https://rust-lang.github.io/rust-clippy/rust-1.74.0/index.html#/explicit_deref_methods
 - - easier and clearer just to dereference
 - https://rust-lang.github.io/rust-clippy/rust-1.74.0/index.html#/missing_const_for_fn
 - - non-const functions prevent callers from being const