growth-ring: Add remaining lint checks
Removed a couple of index dereferences and unwraps.

Unfortunately, there's one spot where it isn't easy to add the lint
check, so there is a block that disables it. It's still a lot better
having it in most of the project and not in this block.
(search for TODO: not allowed for the spot).

Rewriting the loop to be more idiomatic is probably the right move here,
but that's outside the scope of this fix.