Write panics to the log
The orchestrator does not save the standard output of this executable,
it expects everything to be logged. We have two options here, one is to
just write the panics to the log (which is now a required option) and
the other is to redirect stderr into the log. I went with the former,
since I want to see the "starting up" message if I run it interactively.

This could be a little better -- we could both write it to the log and
then render the panic to stderr, but that doesn't seem possible with the
log_panic crate.