Delay block_in_place calls
It's safe to call `block_in_place` if we're not in an async runtime, or
even to do it recursively. The docs suggest you defer calling it until
you know you're about to actually block. This removes the calls from
several locations and puts them all in the one blocking call we have --
the one that reads from the cache.

This didn't make a big difference, but it does eliminate the extra
overhead of moving things over to another thread prematurely.

Will attach flamegraphs of 1,000 inserts.