Don't await for the test results
We can't call await on the future passed from the setup function, but
the client has to be created in the same runtime context as the tests,
so we just block on it in a synchronous context before running the
benchmarks. Unfortunately, we still need to clone and move the client
because it can't go out of scope during the async runtime, but the clone
costs are paid outside of the timing loop.