Fixup TrieHash
TrieHash was previously Copy, but it shouldn't be, as we should be
careful how often we copy hashes around. They are 256 bits, which is not
a single operation.

Additionally, if we use the same type as Sha3, we can avoid copying the
return from `finalize` and just move the result into the TrieHash.

Implemented some additional helpers for TrieHash as well.