Add child_hashes to BranchNode
We have to store these. We might be a little smarter about not
requiring rehashing in some cases, but right now we aggressively
rehash whenever setting the child address.

TODO: The logic for invalidate_hashes in merkle is now broken,
since what we need to invalidate now is the parent's child_hashes.

The HashedNodeStore should be tracking new nodes and what needs
rehashing so that when we transform the NodeStore into an immutable one,
we recompute the hashes before finally writing them to the underlying
linear store.