Hashes use the full key, not the partial_path
There was some effort spent in making this efficient. One snag was that
the digest functions require a `&[u8]` so we can't just use the
iterators. This is unfortunate and we should consider adding methods
that accept iterators instead of slices and adding these to the sha3
crate.

I'm still not convinced this is done the same way as merkledb, but it
should be very easy to change the way it is hashed as the only function
you need to inspect is `HashedNodeStore::hash_internal`.