Add lifetimes to proposals
The underlying database must live longer than the proposal. These lifetime changes enforce that.

Additionally, we need interior mutability for the revision manager. Added a RwLock around it so that it can be locked read-write momentarily during a commit, but read read-only for most other operations.

Gutted the old commit code since we need to follow the spec.