Implement more API methods
Uncommented a lot of the original API code, as it now works.

Fixed the last remaining benchmark. Out of the box, there is less cache so it's slower than the old implementation. With a larger cache (20k nodes), it outperforms the old code (6s for 1M rows vs 84s).

Also fixed propose so it doesn't require a mutable DB.