Expose graph option for fwdctl
Dump was already dumping the keys and values, so we should keep
that, but added in the option of calling the dump() method that
produces a digraph of the trie.

Additional changes: hashes and binary values are truncated to 6
characters, followed by an ellipsis, and the graphs are rotated
so they go left-right. This makes the graphs a lot prettier.

Partial paths can still be long but truncating them seemed like
a bad idea, so they are the only long string in the boxes now.