Add a child option of Arc<Node>
First step in deferring allocation of space is to allow hashed nodes
without addresses to exist. This adds support for that. Note that
these have to be reference counted because child proposals will need
to make deep copies, and need to keep the modified nodes in memory.

This isn't perfect, since if you have proposals A->B->C and A puts
things in memory that are untouched by B and C, the space used for
the in-memory versions in A will not be freed until B and C aren't
referenced any more because they are committed or dropped.

The alternative to this is to have the Child contain a mutable
reference with a lock since commiting A can happen while B or C are
being traversed.