Don't use smallvec for the value
The caller passes in a box, and we then convert it to a smallvec.
Usually this results in the box being converted to a slice, then
a new SmallVec is constructed from that, which allocates another
Box if the value was larger than 16 bytes.

This code is also suspect in that it might be leaking. The difference
between two test cases is only on the size of the value. When the
value was larger than 16 bytes, the memory was growing, otherwise not.

So not only is this faster, but it might avoid the leak we see with
values over 16 bytes.