Remove generic for NodeStore<_, FileBacked>
This previously used S which implemented WritableStorage. However, these
methods are never used unless the backing store is FileBacked, since
they perform write operations to the disk. Removed the generic to make
the next steps easier.

This is probably incomplete in that there are other uses of
WritableStorage; this will need revisiting when we support nodes in
memory.