Include a mode for cache metrics.
There are three modes, 'read' which is for immutable nodestores, 'write' which is for mutable ones, and 'open' which happens during opening of a database and is going to be a miss.

From an end-user POV, these cache stats will correlate to read operations on committed revisions and reads that happen during proposals.