Remove MerkleError and DbError
These errors are almost always std::io::Error. The one case that wasn't
was supposed to be a ProofError, so this eliminates a source of
incorrect error types.

Additionally enabled one more merkle unit test. The test had to change
a little since now an empty trie hash a None hash instead of a default
value.