Fix compiler errors when not using ethhash
The problem here is still some grey areas around when TrieHash is really
required and when HashType is supposed to be used. Added some clippy
allows for useless_conversions which the compiler should optimize away
anyway to work around this, but a better approach might be to use
another method on TrieHash to create a HashType from it rather than just
calling into which can give the clippy warning.