Implement ethhash
This adds full ethereum compatibility including hashing of account
subtrees. The toughest areas are around accounts, since we need a node
that is a fake root node in some cases, and that can change -- an
account branch that previously had one child could now have more than
one which means it has to rehash something that didn't change.

This means that hash_helper and friends can now fail, as they do I/O.
The propogation of the error back to the callers now passes all the way
back through the API.

CI now tests with logger and ethhash features enabled, and again with
them disabled.