Add some additional open checks
If the database was created with a specific ethhash value, this prevents
someone from opening it with a different value.

Also, changing the sizes of the freelists is a database incompatible
change, so we get the hash of that and store it in the header. This
again prevents someone from opening an incompatible database.