Windows OS support
Heard that we might need this, as some people want to run avalanchego
on windows.

This uses the windows equivalent functions in places where we used
unix specific functions.

`grpc-testtool` is no longer compiled at the top level. There are a
number of unix-specific calls there, and we're currently not using
it anyway.