fix: Fix empty hash values
When the trie is empty, be consistent about returning the same thing.
This fix makes the "same thing" be:

 - If ethhash is enabled, always return the ethereum-compatible empty
   hash
 - If ethhash is disabled, return None.

If ethhash is disabled, the caller can easily map None to whatever empty
hash they want. In hindsight, probably ethhash should return None as
well for empty hashes, but that would require the callers to know what
the magic hash is for ethhash.