fix: Reimplement
Took a slightly different approach, since the previous diff didn't
use the optimized node iterator, and didn't even check for it.

Other changes:
 - Renamed children_have_same_hash() to child_identical()
 - Added documentation around more complicated functions
 - Added additional tests to verify that fewer nodes are read when
   diffing tries with hashes already computed
 - Installed a test metrics recorder so metrics can be checked at
   runtime

Note that if we add additional tests that want to use this metrics
recorder, we would need to put this into a utility method and save it in
case some other test initializes the metrics.