test(ethhash): convert ethhash test to fuzz test for ethhash compatibility (#956)
Update ethhash compatibility test to a fuzz test.

Follow the style from https://github.com/ava-labs/firewood/pull/930 to
define a state machine that randomly selects an operation and leaves to
the fuzzer to determine the order of steps to run and the inputs as much
as possible.

Future improvements could be to leave as much to the fuzzer as possible
(ie move from using a random seed to determine inputs to leaving all
inputs to the fuzzer).