fix(storage): Restrict database version identifier
NOTE: This will make the previous database unreadable; but, this would
also happen every time we bumped the patch version.

Problem: Currently, the database version includes both the minor and
patch numbers as well as any other potential values in
`CARGO_PKG_VERSION`. Therefore, whenever the patch version is bumped,
which is expected to happen fairly often, the database becomes
incompatible.

Solution: We're truncating the version included in the db to just the
major version. Therefore, when we eventually release 1.0.0, then will be
a database incompatible upgrade.

Resolves: #958