fix: Report IO errors
Previously, if an error occurred while flushing nodes, it was silently
ignored. To reproduce this, I created a tiny filesystem using:

    sudo mkdir /mnt/tmpfs
    sudo mount -t tmpfs -o size=5k,mode=1777,noatime tmpfs /mnt/tmpfs

I then ran a few fwdctl commands, which passed on main:

    cargo run --bin fwdctl --features logger -- create /mnt/tmpfs/test.db
    target/debug/fwdctl insert --db /mnt/tmpfs/test.db a "$(tr -d '\n' < /etc/passwd)"
    target/debug/fwdctl insert --db /mnt/tmpfs/test.db b "$(tr -d '\n' < /etc/passwd)"
    target/debug/fwdctl insert --db /mnt/tmpfs/test.db c "$(tr -d '\n' < /etc/passwd)"

When running this after these fixes, the following is produced from the
final insert, which fills the disk:

    Error: FileIO(FileIoError { inner: Os { code: 28, kind: StorageFull, message: "No space left on device" }, filename: Some("/mnt/tmpfs/test.db"), offset: 8544, context: Some("write failure") })

This fix restructures the pinned buffer pool to add an offset, which is
only used for error reporting. This offset is also used as a flag
indicating the entry is in use rather than using a bool.