perf: Cache the latest view
Because of the usage pattern for coreth, there is contention between
committing and getting values. Coreth attempts to commit the latest
version asynchronously while fetching values from it.

This is fixed by adding a single cached revision. In the normal case,
there will be many reads on some revision, followed by creating a
proposal, followed by an asynchronous commit. During this commit, there
will be additional reads, so we save the view just before committing.