fix: publish firewood-macros
The macros crate was not being published due to the missing stanza. Need
to figure out a way to make this automatic in the future.

I also added a note about how the crates in the publish workflow must be
sorted in reverse topological order so that the dependencies are
published before the packages that depend on them.