chore: share workspace metadata and packages
Many of the package fields can be declared at the workspace level and
then reused by the individual packages. This reduces duplication and
drift.

Additionally, moving the workspace packages to be listed as workspace
dependencies instead of strewn throughout the multiple Cargo.toml files
makes it to bump version numbers across the workspace.
`cargo workspace-version` will be able to automatically bump everything
as opposed to before only partially updating the workspace.