fix: logger macros causing linting warnings (#1021)
Previously, the logger macros were empty when logging was disabled. This caused no code to be generated which resulted in warnings from clippy around dead code.

This change bypasses those warnings by generating code that lexically uses the parameters passed in however at runtime will never be executed and will be elided by the compiler when optimizations are enabled and logging is disabled.