feat(ffi): Add gauges to metrics reporter
Gauges weren't being reported via the homegrown HTTP endpoint we used
for ffi, now they are. This is particularly important since we now have
a few useful gauges in firewood.

The surrounding HTTP structure might be going away, but the underlying
`stats` call is likely to stay, so this enhancement is probably still
useful.

Also took the opportunity to reduce memory allocations here, and write
directly to strings instead of converting back and forth between
string/bytes/string with all the utf8 conversions.

Histograms are still not supported, but we don't have any, and aren't
likely to add any anytime soon.