feat(delayed-persist): 2.1: Unpersisted deletions
Nodes on the delete list might be unpersisted. This can't actually
happen yet, but support for it is there. This is half of making branch
children unpersisted.

Reaping unpersisted nodes is easy, we just drop them. If they are
persisted, they need to move to the freelist, as they did before.