feat(deferred-persist): UpersistedIterator
An UnpersistedIterator iterates through a NodeStore looking for
unpersisted noded. Unpersisted nodes are only of the type
MaybeUnpersisted that are not yet persisted.

Note that it's impossible to have a persisted parent for a
MaybeUnpersistedNode, so it's relatively easy to find the unpersisted
ones.

There are tests that verify that the nodes are returned in the order
that they will need persisting (that is, depth first).