feat(deferred-allocate): Allocate in parts
The `NodeStore` was the defacto source for allocations, but it required
an owned nodestore and had some borrowing issues in upcoming diffs. This
change uses a `NodeAllocator` struct which just points to the header.

This diff just constructs one of these for compatibility with the old
code. A future PR will create a single allocator for all allocations and
then mutate the header all at once.