feat(deferred-persist): Part 1: unpersisted gauge
Added a gauge to track the number of unpersisted nodes at any time.
These can get used later by the revision manager to decide whether or
not to persist a proposal (among other possible strategies).

It's also helpful to know how many nodes are still in memory that could
be persisted.

When increasing this counter, we do it all at once, but we decrement one
by one (or by the number of entries from the kernel's completion queue).
This is because the writes are slower than the adds, and there is some
integer to floating point math that happens each time the counter is
adjusted.