feat: add `Children<T>` type alias
This commit introduces a new type alias `Children<T>` which is a fixed-size
array of `Option<T>`. This is convenient for all of the places where we
represent the children of a branch node in the trie.

Also included is a method to construct the array in const context, allowing
for easy creation of `Children<T>` instances.