feat: make NodeStore more generic
This commit enhances the `NodeStore` structure to be more generic,
allowing it to work with different storage backends. In a future change,
I have tests that execute a write-commit-read loop using `MemStore`.
However, this does not work without this change as `NodeStore` currently
has many methods that are specific to the `FileStore` implementation.

Passing in an `Arc` is not required as we're not cloning the arc. This
makes it easier to use the `NodeStore` without needing to wrap it in an
`Arc` every time.

A `fork` method is also added to `NodeStore` for easier cloning of the
store into a new mutable instance.