feat: Update RangeProof structure
While implementing the range proof verification, I realized that having
an Option over a byte range is not ideal to work with. This mean that I
needed to do `proof.is_none_or(|p| p.is_empty())` every time I
wanted to use a proof in the range. It was easier to use `proof.is_empty()`
everywhere. The inverse was also true where when I wanted to use a proof,
I first needed to destructure it and then check if it was empty.

Also updated is the method signature and documentation for `verify_range_proof`
to match how it will be used in the future.