fix: Fix broken deserialization of old FreeArea format
As noted in the issue, we were failing to deserialize FreeAreas when
they were encoded in the old serde format. This was because I did not
correctly implement deserializing the varints. Bincode's format is
different from the format `integer-encoding` uses thus the problem when
deserializing.

This fixes the issue by implementing the correct deserialization logic
for FreeArea. Now, it can correctly read the old format.

Fixes: #1146