ci: run CI on macOS
As noted in #1129 and #1135, there are some inconsistencies with builds
for macOS and Linux. This commit adds a macOS runner to the CI workflow
to ensure that the lints and tests pass on macOS as well.

Instead of splitting the matrix to hoist the os separately, I simply
added a new element to the include block. This way the GithHub blocks
on merging Pull Requests won't be confused by the change in matrix values
(it matches on all of them literally).

Fixes #1129