feat: update revision manager error
Previously, the api::Error type had duplicate variants that also covered
the errors from RevisionManagerError. This commit removes the duplicates
and makes updates RevisionManagerError to include the fields necessary
to translate the error to the api::Error type using the existing
variants.

Additionally, the message "commit the parents of this proposal first" is
inaccurate. We also will be unable to commit a proposal when the parent
is already committed if the parent is again not the latest revision. In
this case, the message to commit the parents first doesn't make sense
because they are already committed. We actually need to re-parent the
proposal but currently do not have a mechanism for this via FFI.