fix: Fix race around reading nodes during commit
We called persist_at before th enode was actually persisted to disk in
the io-uring case. What we really want is to keep all the nodes around
in memory while they are being flushed, and then once they are added
to the general cache, we can mark them as fully persisted.

This is a lot better since now nodes will be used directly while they
are being flushed instead of having a disk address that either hasn't
been written to disk or at least isn't yet in the cache.