fix: use `count` instead of `size_hint`
`size_hint` is not guaranteed to be accurate and in some cases where we
use it, it may not be. This can potentially lead to silent bugs due to
incorrect assumptions about the number of elements in an iterator.

I also refactored nibbles_to_eth_compact to be clearer and take
advantage of newer rust features to remove some redundant code.