fix(range-proofs): fix verify of exclusion proofs
We were incorrectly checking the length of the last node's full path
against the length of the key being proven. This meant that if the
last node's full path was the same _length_ as the key being proven, but
the key was different, we would incorrectly verify the proof by expecting
a value when there should be none.