refactor: unify root storage using Child enum
There were two fields representing a root, the address and the hash.
They were both options, which means that it is possible for one to be
None and the other to be Some, which would be a logic error. This PR
makes that impossible.

Thought hard about whether or not to use Child, as it's possible to use
something simpler than that, but based on the fact that we're going to
add deferred persistence, this is probably the right structure. The
owned Node variant of Child will never be used, but this problem is
systemic and perhaps worthy of another PR and is much more difficult.

Summary of changes:

- Replace separate root_hash and root fields with single Child-based root
- Update all root access methods to use Child::as_maybe_persisted_node()
- Fix test compilation errors by properly chaining method calls
- Simplify root persistence logic using Child::persisted_address()