feat: add newtype for PathComponent
The new `PathComponent` type is introduced to encapsulate the possible
values of a single component in the Trie path. If `branch_factor_256` is
not enabled, the default, `PathComponent` wraps `U4`, which was added in
the previous pull request. This type is only value for values 0-15,
inclusive. If `branch_factor_256` is enabled, `PathComponent` wraps a
`u8`, which can represent values 0-255, inclusive.

This change contributes to the resolution of #1230, but does not fully
resolve it until all usages of `Path` are replaced with future changes.