fix: EINTR during iouring calls
golang does not guarantee signals are using SA_RESTART, so any system
call on a "slow device" might cause the EINTR error. We saw this in the
error in issue 1347.

This doesn't resolve that issue completely because IO errors can occur
during these writes. That will be fixed under a separate issue.

My research indicated that re-calling the system call is the right thing
to do here.