feat(ffi): `Database.Close()` guarantees proposals committed or freed (#1349)
Explicit freeing of proposals propagated through `libevm` (i.e. `geth`)
plumbing has proven difficult when not being committed as they are
simply dropped for the GC to collect. Furthermore, strict ordering of
calls to `Proposal.Drop()` (or `Commit()`) before `Database.Close()` is
required to avoid segfaults. This PR implements a fix for both issues:

1. All new `Proposal`s have a GC finalizer attached, which calls
`Drop()`. This is safe because it is a no-op if called twice or after a
call to `Commit()`.
2. The `Database` has a `sync.WaitGroup` introduced, which tracks all
outstanding proposals. Calls to `Commit()` / `Drop()` decrement the
group counter (only once per `Proposal`).
3. `Database.Close()` waits on the `WaitGroup` before freeing its own
handle, avoiding segfaults.

Assuming that all calls to `Database.Propose()` and `Proposal.Propose()`
occur _before_ the call to `Database.Close()` then this is a correct
usage of `sync.WaitGroup`'s documented requirement for ordering of calls
to `Add()` and `Wait()`.

An integration test demonstrates blocking and eventual return of
`Database.Close()`, specifically due to the unreachability of
un-dropped, un-committed `Proposal`s, resulting in their finalizers
decrementing the `WaitGroup`.

---------

Signed-off-by: Arran Schlosberg <519948+ARR4N@users.noreply.github.com>
Co-authored-by: Austin Larson <78000745+alarso16@users.noreply.github.com>