fix: fix `giant_node` test
The previous work to defer allocating nodes until commit time
had the side effect of allowing nodes to exceed the maximum allowed size
at hash time. This defers the error until commit time, which is not
ideal.

This change updates the test to confirm that the allocator still errors
when trying to allocate a node that is too large. However, we would
prefer to do this earlier but without serializing the node multiple
times.

As part of this change, I created a helper type that the test can use
to verify the exact error was returned instead of comparing against
the string representation of the error. As a part of that, I discovered
that there was some vestiges of old code where the allocator would
potentially return an area larger than necessary when allocating from
the free list. However, that functionality was removed in #1217; thus,
I have simplified the remaining code to reflect that which also
simplified my `giant_node` test changes.

I added `DisplayTruncatedHex` to the `Debug` representations of the
leaf and branch nodes to truncate the node values when printing debug
info. This was necessary before fixing the `giant_node` test because
the debug output included the full contents of the node, which would
flood the terminal when the test failed and cause it to hang.

Closes: #1054