fix: mark node as allocated before serializing parent
An issue was found where during serialization of a node, it would panic
indicating that a child node was not allocated and therefore could not
serialize the parent node. This occurred after #1389 because we began
serializing nodes in batches in order to prevent corrupting the free-
list after a failed write. Previously, we would serialize a node, write
it to storage, and then mark it as allocated in one step. Now we serialize
and allocate the node and add the node to a batch before moving onto the
next node. Like before, we only marked the node as allocated after
writing it to disk. This means that if the batch contained child and
parent, the parent node fails to serialize because the child node has
been allocated but the parent is unaware.

This change refactors the allocation logic to mark nodes as allocated
before batching. This ensures that if its parent is also in the batch,
it can successfully serialize. This linear address itself is not read
from by viewers of the `Allocated` state because the node is not
considered committed until the entire write has completed (i.e., after
the root and header are written). This means this change does not
introduce any early reads of uncommitted data.

In the course of writing tests, I discovered that the arena allocator
returns the number of bytes by the arena, not the number of bytes
occupied by items within the arena. This meant that the batching logic
should have always generated a batch of 1 element. To fix this, I track
the allocated length of each area. This successfully triggered the bug
I discovered; however, this also means the bug might not be the source
of the original panic.