refactor(rootstore): own nodestore logic (#1495)
As mentioned in #1470, there exists the following circular dependency: 

```mermaid
stateDiagram-v2
    s1: Revision Manager
    s2: RootStore

    s1 --> s2: Wants historical revision X
    s2 --> s1: Wants latest revision
```

This PR eliminates the circular dependency above by giving `RootStore` a
reference to the file backend, which allows it to construct revisions
without having to rely on the storage of another `NodeStore`
(previously, we needed the latest `NodeStore` whose storage we would
clone).

This PR does the following:
- `NodeStore` construction with a specified root hash + address now
takes in a storage reference rather than another `NodeStore.
- Revision construction is now handled by `RootStore` instead of the
RevisionManager
- Moves revision caching from the RevisionManager to `RootStore`
- Adds tests for `RootStore` caching logic

Note: there exists an earlier version of this PR (#1480). However,
https://github.com/ava-labs/firewood/pull/1480#discussion_r2566078111
pointed out that having to pass in a `NodeStore` to
`NodeStore::with_root()` seemed odd. After some thought, I agree with
this point and found that passing in a storage reference instead removed
the need to have to share revisions between the `RevisionManager` and
`RootStore`. Hence, I went with creating a new PR rather than clobbering
the existing one.